Fix for APP009 of the compliance test by Pan Jie
git-svn-id: http://partuza.googlecode.com/svn/trunk@170 329b45d6-434c-0410-81e3-f98314b225b0

+++ Shindig/PartuzaDbFetcher.php
@@ -216,10 +216,16 @@ class PartuzaDbFetcher {
   public function deleteAppData($person_id, $key, $app_id) {
     $this->checkDb();
     $person_id = mysqli_real_escape_string($this->db, $person_id);
-    $key = mysqli_real_escape_string($this->db, $key);
     $app_id = mysqli_real_escape_string($this->db, $app_id);
-    if (! @mysqli_query($this->db, "delete from application_settings where application_id = $app_id and person_id = $person_id and name = '$key'")) {
-      return false;
+    if ($key == '*') {
+      if (! @mysqli_query($this->db, "delete from application_settings where application_id = $app_id and person_id = $person_id")) {
+        return false;
+      }
+    } else {
+      $key = mysqli_real_escape_string($this->db, $key);
+      if (! @mysqli_query($this->db, "delete from application_settings where application_id = $app_id and person_id = $person_id and name = '$key'")) {
+        return false;
+      }
     }
     return true;
   }
